refactor(combat): 重构战斗格子系统，新增MapCell封装地图格子逻辑
1. 新增MapCell纯数据类封装格子坐标与占用关系
2. 替换gridOccupancy列表为MapCell二维数组管理格子状态
3. 统一回合制数值类型为int，移除float转round的冗余代码
4. 新增格子持续效果CellEffect与相关操作逻辑
5. 重构技能CD、召唤时长相关逻辑为回合数管理
6. 新增棋子行动速度属性与回合排序优化

+++ Assets/Resources/Scripts/Combat/Actions/AddBuffAction.cs
@@ -4,9 +4,9 @@ public class AddBuffAction : ChessAction
     public int CasterId;
     public int SkillId;
     public int BuffId;
-    public float LastRounds;
+    public int LastRounds;
 
-    public AddBuffAction(int sourceId, int tick, int casterId, int skillId, int buffId, float lastRounds)
+    public AddBuffAction(int sourceId, int tick, int casterId, int skillId, int buffId, int lastRounds)
         : base(sourceId, tick)
     {
         CasterId = casterId;
